+++ public/js/admin/adminCtrl.js
@@ -1,6 +1,5 @@
 angular.module("app")
     .controller("adminCtrl", function($scope, adminSvc, $state, sessions) {
-        console.log("sessions", sessions);
 
         $scope.adminUser = true;
         $scope.manageProducts = true;
@@ -27,11 +26,14 @@ angular.module("app")
         $scope.addProduct = function(product) {
             adminSvc.addNewProduct(product);
             callGetProducts();
+            $scope.product = " ";
         };
 
-        $scope.deleteProduct = function(id) {
-            adminSvc.destroyProduct(id);
-            callGetProducts();
+        $scope.deleteProduct = function(id, product) {
+            if (confirm("Are you sure you want to delete the " + product + "?")) {
+                adminSvc.destroyProduct(id);
+                callGetProducts();
+            }
         };
 
         $scope.updateProduct = function(id, name, price, img, model, summary, condition, type) {
@@ -75,6 +77,7 @@ angular.module("app")
         $scope.addAdmin = function(admin) {
             adminSvc.addAdmin(admin);
             callGetAdmins();
+            $scope.admin = " ";
         };
 
         $scope.updateAdmin = function(id, admin) {
@@ -82,9 +85,11 @@ angular.module("app")
             callGetAdmins();
         };
 
-        $scope.deleteAdmin = function(id) {
-            adminSvc.deleteAdmin(id);
-            callGetAdmins();
+        $scope.deleteAdmin = function(id, name) {
+            if (confirm("Are you sure you want to remove " + name + " from the admin list?")) {
+                adminSvc.deleteAdmin(id);
+                callGetAdmins();
+            }
         };
 
         var callGetLessons = function() {
@@ -98,11 +103,14 @@ angular.module("app")
         $scope.newLesson = function(lesson) {
             adminSvc.addLesson(lesson);
             callGetLessons();
+            $scope.lesson = " ";
         };
 
-        $scope.deleteLesson = function(lessonID) {
-            adminSvc.deleteLesson(lessonID);
-            callGetLessons();
+        $scope.deleteLesson = function(lessonID, student) {
+            if (confirm("Are you sure you want to delete lesson for " + student + "?")) {
+                adminSvc.deleteLesson(lessonID);
+                callGetLessons();
+            }
         };
 
         $scope.updateLesson = function(id, lesson) {
@@ -111,26 +119,6 @@ angular.module("app")
         };
 
         // *************************** calendar ****************
-
-        var obj = {
-            year: 2016,
-            month: 4,
-            day: 3
-        };
-
-        // var events = [{
-        //     title: "Mom's Birthday",
-        //     start: new Date(obj.year, obj.month, obj.day)
-        // }, {
-        //     title: "Party Time",
-        //     start: new Date(2016, 4, 12),
-        //     end: new Date(2016, 4, 16)
-        // }, {
-        //     title: "Mom's Birthday",
-        //     start: new Date(obj.year, obj.month, obj.day)
-        // }];
-        // console.log("events", events);
-
         events = [];
         $scope.eventSources = [events];
         var callGetStudioSessions = function() {
@@ -146,10 +134,6 @@ angular.module("app")
                 });
 
         };
-
-
-
-        // $scope.eventSources = [events];
         $scope.calOptions = {
             editable: true,
             header: {
@@ -167,17 +151,14 @@ angular.module("app")
         };
 
         $scope.deleteStudioSession = function(id) {
-            adminSvc.deleteStudioSession(id);
-            callGetStudioSessions();
+            if (confirm("Are you sure you want to delete this Studio Session?")) {
+                adminSvc.deleteStudioSession(id);
+                callGetStudioSessions();
+            }
         };
 
         $scope.updateStudioSession = function(id, session) {
             adminSvc.updateStudioSession(id, session);
             callGetStudioSessions();
         };
-
-
-
-
-
     });
